+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `IHazGitHubActions`
 - Add `TestResultUtil` / `TestResultUtil.Markdown`
 - Add `TrxExtension`
+### Tests
+- Update `Tests` with fail test, ignore and pass.
 
 ## [1.5.0] / 2023-05-31
 ### Updated